+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] -
+### Fixed
+- Backup related tasks are not scheduled twice.
+- Backup retention policy task is scheduled.
+
 ## [1.5.1] - 2022-04-28
 ### Fixed
 - Bug in new release deployment procedure.